Moved spatial function from main.utils to main.spatial_utils
