+++ boranga/components/occurrence/utils.py
@@ -13,7 +13,7 @@ from django.db import transaction
 from django.db.models import Q
 from django.utils import timezone
 
-from boranga.components.main.utils import feature_json_to_geosgeometry
+from boranga.components.main.spatial_utils import feature_json_to_geosgeometry
 from boranga.components.occurrence.email import (
     send_external_submit_email_notification,
     send_submit_email_notification,
